Fix wasm32 build
