+++ vm/src/stdlib/thread.rs
@@ -1,5 +1,5 @@
 //! Implementation of the _thread module
-#[cfg_attr(target_os = "wasi", allow(unused_imports))]
+#[cfg_attr(target_arch = "wasm32", allow(unused_imports))]
 pub(crate) use _thread::{make_module, RawRMutex};
 
 #[pymodule]
